+++ service/src/main/kotlin/io/provenance/explorer/domain/extensions/MessageTranslationExtensions.kt
@@ -77,6 +77,9 @@ fun Message.toObjectNodeNonTxMsg(protoPrinter: JsonFormat.Printer, fieldNames: L
             node
         }
 
+fun Message.toObjectNodePrint(protoPrinter: JsonFormat.Printer) =
+    OBJECT_MAPPER.readTree(protoPrinter.print(this))
+
 fun Message.toObjectNodeMAddressValues(protoPrinter: JsonFormat.Printer, fieldNames: List<String>) =
     OBJECT_MAPPER.readTree(protoPrinter.print(this))
         .let { node ->