+++ service/src/main/kotlin/io/provenance/explorer/domain/models/explorer/ParamModels.kt
@@ -1,6 +1,6 @@
 package io.provenance.explorer.domain.models.explorer
 
-import com.google.protobuf.ProtocolStringList
+import com.fasterxml.jackson.databind.JsonNode
 
 data class Params(
     val cosmos: CosmosParams,
@@ -8,26 +8,14 @@ data class Params(
 )
 
 data class CosmosParams(
-    val auth: AuthParams,
-    val bank: BankParams,
+    val auth: JsonNode,
+    val bank: JsonNode,
     val dist: DistParams,
     val gov: GovParams,
     val mint: MintParams,
     val slashing: SlashingParams,
-    val staking: StakingParams,
+    val staking: JsonNode,
     val ibc: IBCParams,
-)
-
-data class AuthParams(
-    val max_memo_characters: Long,
-    val txSigLimit: Long,
-    val txSizeCostPerByte: Long,
-    val sigVerifyCostEd25519: Long,
-    val sigVerifyCostSecp256k1: Long,
-)
-
-data class BankParams(
-    val defaultSendEnabled: Boolean
 )
 
 data class DistParams(
@@ -38,29 +26,15 @@ data class DistParams(
 )
 
 data class GovParams(
-    val voting: VotingParams,
+    val voting: JsonNode,
     val tallying: TallyingParams,
-    val deposit: DepositParams,
-)
-
-data class VotingParams(
-    val votingPeriod: Long,
+    val deposit: JsonNode,
 )
 
 data class TallyingParams(
     val quorum: String,
     val threshold: String,
     val vetoThreshold: String
-)
-
-data class DepositParams(
-    val minDeposit: MinDeposit,
-    val maxDepositPeriod: Long,
-)
-
-data class MinDeposit(
-    val denom: String,
-    val amount: String,
 )
 
 data class MintParams(
@@ -75,53 +49,19 @@ data class MintParams(
 data class SlashingParams(
     val signedBlocksWindow: Long,
     val minSignedPerWindow: String,
-    val downtimeJailDuration: Long,
+    val downtimeJailDuration: String,
     val slashFractionDoubleSign: String,
     val slashFractionDowntime: String,
 )
 
-data class StakingParams(
-    val unbondingTime: Long,
-    val maxValidators: Int,
-    val maxEntries: Int,
-    val historicalEntries: Int,
-    val bondDenom: String,
-)
-
 data class IBCParams(
-    val transfer: TransferParams,
-    val client: ClientParams,
-)
-
-data class TransferParams(
-    val sendEnabled: Boolean,
-    val receiveEnabled: Boolean,
-)
-
-data class ClientParams(
-    val allowedClientsList: ProtocolStringList,
+    val transfer: JsonNode,
+    val client: JsonNode,
 )
 
 data class ProvParams(
-    val attribute: AttributeParams,
-    val marker: MarkerParams,
-    val metadata: String,
-    val name: NameParams,
-)
-
-data class AttributeParams(
-    val maxValueLength: Int,
-)
-
-data class MarkerParams(
-    val maxTotalSupply: Long,
-    val enableGovernance: Boolean,
-    val unrestrictedDenomRegex: String,
-)
-
-data class NameParams(
-    val maxSegmentLength: Int,
-    val minSegmentLength: Int,
-    val maxNameLevels: Int,
-    val allowUnrestrictedNames: Boolean,
+    val attribute: JsonNode,
+    val marker: JsonNode,
+    val metadata: JsonNode,
+    val name: JsonNode,
 )