+++ service/src/main/kotlin/io/provenance/explorer/domain/models/explorer/ValidatorModels.kt
@@ -61,6 +61,11 @@ data class Delegation(
     val endTime: DateTime?
 )
 
+data class UnpaginatedDelegation(
+    val records: List<Delegation>,
+    val rollupTotals: Map<String, CoinStr>
+)
+
 data class ValidatorCommission(
     val bondedTokens: CountStrTotal,
     val selfBonded: CountStrTotal,